// autogpt_platform/frontend/src/app/api/__generated__/models/node.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { NodeInputDefault } from "./nodeInputDefault";
import type { NodeMetadata } from "./nodeMetadata";
import type { Link } from "./link";

export interface Node {
  id?: string;
  block_id: string;
  input_default?: NodeInputDefault;
  metadata?: NodeMetadata;
  input_links?: Link[];
  output_links?: Link[];
}